% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords-translate.R
\name{coords_point_translate_wkt}
\alias{coords_point_translate_wkt}
\title{Parse coordinates into well-known formats}
\usage{
coords_point_translate_wkt(x, y, z = NA, m = NA, precision = 16, trim = TRUE)
}
\arguments{
\item{x, y, z, m}{Vectors of coordinate values}

\item{precision}{The rounding precision to use when writing
(number of decimal places).}

\item{trim}{Trim unnecessary zeroes in the output?}
}
\value{
\verb{*_translate_wkt()} returns a character vector of
well-known text; \verb{*_translate_wkb()} returns a list
of raw vectors, and \verb{*_translate_wksxp()} returns an unclassed
list of \code{\link[=wksxp]{wksxp()}} geometries.
}
\description{
These functions provide the reverse function of \code{\link[=wkt_coords]{wkt_coords()}}
and company: they parse vectors of coordinate values into well-known
formats.
}
\examples{
coords_point_translate_wkt

}
