% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crc.R
\name{crc}
\alias{crc}
\alias{as_crc}
\alias{as_crc.wk_crc}
\alias{as_crc.matrix}
\alias{as_crc.data.frame}
\title{2D Circle Vectors}
\usage{
crc(x = double(), y = double(), r = double(), crs = wk_crs_auto())

as_crc(x, ...)

\method{as_crc}{wk_crc}(x, ...)

\method{as_crc}{matrix}(x, ..., crs = NULL)

\method{as_crc}{data.frame}(x, ..., crs = NULL)
}
\arguments{
\item{x, y}{Coordinates of the center}

\item{r}{Circle radius}

\item{crs}{A value to be propagated as the CRS for this vector.}

\item{...}{Extra arguments passed to \code{as_crc()}.}
}
\value{
A vector along the recycled length of bounds.
}
\description{
2D Circle Vectors
}
\examples{
crc(1, 2, 3)

}
