% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crc.R
\name{crc_x}
\alias{crc_x}
\alias{crc_y}
\alias{crc_center}
\alias{crc_r}
\title{Circle accessors}
\usage{
crc_x(x)

crc_y(x)

crc_center(x)

crc_r(x)
}
\arguments{
\item{x}{A \code{\link[=crc]{crc()}} vector}
}
\value{
Components of the \code{\link[=crc]{crc()}} vector
}
\description{
Circle accessors
}
\examples{
x <- crc(1, 2, r = 3)
crc_x(x)
crc_y(x)
crc_r(x)
crc_center(x)

}
