% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd-subset.R
\name{grd_index.wk_grd_xy}
\alias{grd_index.wk_grd_xy}
\title{#' @export
grd_index.wk_grd_rct <- function(grid, point, ..., snap = round) {
s <- grd_summary(grid)
point <- unclass(as_xy(point))
i <- if (s$height == -Inf) rep(NA_real_, length(point$x)) else (point$x - s$xmin) / s$dx
j <- if (s$width == -Inf) rep(NA_real_, length(point$x)) else (s$ymax - point$y) / s$dy
list(i = snap(i + 1L), j = snap(j + 1L))
}}
\usage{
\method{grd_index}{wk_grd_xy}(grid, point, ..., snap = round)
}
\description{
#' @export
grd_index.wk_grd_rct <- function(grid, point, ..., snap = round) {
s <- grd_summary(grid)
point <- unclass(as_xy(point))
i <- if (s$height == -Inf) rep(NA_real_, length(point$x)) else (point$x - s$xmin) / s$dx
j <- if (s$width == -Inf) rep(NA_real_, length(point$x)) else (s$ymax - point$y) / s$dy
list(i = snap(i + 1L), j = snap(j + 1L))
}
}
