% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd-overview.R
\name{grd_overview}
\alias{grd_overview}
\alias{grd_overview.grd_rct}
\alias{grd_overview_summary}
\title{Compute overview grid specifications}
\usage{
grd_overview(grid, level, ...)

\method{grd_overview}{grd_rct}(grid, level, ...)

grd_overview_summary(grid, levels = NULL)
}
\arguments{
\item{grid}{A \code{\link[=grd_xy]{grd_xy()}}, \code{\link[=grd_rct]{grd_rct()}}, or other object
implementing \verb{grd_*()} methods.}

\item{level}{An integer describing the overview level. This is related to
the \code{step} value by a power of 2 (i.e., a level of \code{1} indicates a step of
\code{2}, a level of \code{2} indicates a step of \code{4}, etc.).}

\item{...}{Passed to S3 methods}

\item{levels}{A vector of \code{level} values or \code{NULL} to use a sequence from
0 to the level that would result in a 1 x 1 grid.}
}
\value{
A \code{\link[=grd]{grd()}}
}
\description{
Compute overview grid specifications
}
\examples{
grid <- grd_rct(volcano)
grd_overview_summary(grid)
grd_overview(grid, 1)

}
