% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd-resample.R
\name{grd_resample_nearest}
\alias{grd_resample_nearest}
\title{Resample grid objects}
\usage{
grd_resample_nearest(grid, template, trans = wk_affine_identity(), ...)
}
\arguments{
\item{grid}{A \code{\link[=grd_xy]{grd_xy()}}, \code{\link[=grd_rct]{grd_rct()}}, or other object
implementing \verb{grd_*()} methods.}

\item{template}{A grid whose \code{bbox}, \code{nrow()}, and \code{ncol()} will be used
to determine the \code{bbox}, \code{nrow()}, and \code{ncol()} of the output.}

\item{trans}{An external pointer to a wk_trans object}

\item{...}{Passed to S3 methods}
}
\value{
A grid with the same class, \code{bbox}, \code{nrow()}, and \code{ncol()} as
\code{template}.
}
\description{
Resample grid objects
}
