% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd.R
\name{grd_summary}
\alias{grd_summary}
\alias{grd_data_order}
\title{Grid information}
\usage{
grd_summary(grid)

grd_data_order(grid_data)
}
\arguments{
\item{grid}{A \code{\link[=grd_xy]{grd_xy()}}, \code{\link[=grd_rct]{grd_rct()}}, or other object
implementing \verb{grd_*()} methods.}

\item{grid_data}{The \verb{$data} member of a \code{\link[=grd_xy]{grd_xy()}}, \code{\link[=grd_rct]{grd_rct()}},
or other object implementing \verb{grd_*()} methods.}
}
\value{
\itemize{
\item \code{grd_summary()} returns a \code{list()} with components
\code{xmin}, \code{ymin}, \code{xmax}, \code{ymax},
\code{nx}, \code{ny}, \code{dx}, \code{dy}, \code{width}, and \code{height}.
\item \code{grd_data_order()} returns \code{c("y", "x")} for
data with a column-major internal ordering and
\code{c("x", "y")} for data with a row-major internal
ordering. Both 'x' and 'y' can be modified with
a negative sign to indicate right-to-left
or bottom-to-top ordering, respectively.
}
}
\description{
Grid information
}
\examples{
grd_summary(grd(nx = 3, ny = 2))

}
