% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-wkt.R
\name{handle_wkt_without_vector_size}
\alias{handle_wkt_without_vector_size}
\title{Test handlers for handling of unknown size vectors}
\usage{
handle_wkt_without_vector_size(handleable, handler)
}
\arguments{
\item{handleable}{A geometry vector (e.g., \code{\link[=wkb]{wkb()}}, \code{\link[=wkt]{wkt()}}, \code{\link[=xy]{xy()}},
\code{\link[=rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[=wk_handle]{wk_handle()}} is defined.}

\item{handler}{A \link[=wk_handle]{wk_handler} object.}
}
\description{
Test handlers for handling of unknown size vectors
}
\examples{
handle_wkt_without_vector_size(wkt(), wk_vector_meta_handler())

}
