% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{new_grd}
\alias{new_grd}
\title{S3 details for grid objects}
\usage{
new_grd(x, subclass = character())
}
\arguments{
\item{x}{A \code{\link[=grd]{grd()}}}

\item{subclass}{An optional subclass.}
}
\description{
S3 details for grid objects
}
