% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wk-crc.R
\name{new_wk_crc}
\alias{new_wk_crc}
\title{S3 details for crc objects}
\usage{
new_wk_crc(x = list(x = double(), y = double(), r = double()), crs = NULL)
}
\arguments{
\item{x}{A \code{\link[=crc]{crc()}}}

\item{crs}{A value to be propagated as the CRS for this vector.}
}
\description{
S3 details for crc objects
}
