% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{new_wk_grid}
\alias{new_wk_grid}
\title{S3 details for grid objects}
\usage{
new_wk_grid(x, subclass = character())
}
\arguments{
\item{x}{A \code{\link[=wk_grid]{wk_grid()}}}

\item{subclass}{An optional subclass.}
}
\description{
S3 details for grid objects
}
