% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct.R
\name{new_wk_rct}
\alias{new_wk_rct}
\title{S3 details for rct objects}
\usage{
new_wk_rct(
  x = list(xmin = double(), ymin = double(), xmax = double(), ymax = double()),
  crs = NULL,
  geodesic = NULL
)
}
\arguments{
\item{x}{A \code{\link[=rct]{rct()}}}

\item{crs}{A value to be propagated as the CRS for this vector.}

\item{geodesic}{\code{TRUE} if edges must be interpolated as geodesics when
coordinates are spherical, \code{FALSE} otherwise.}
}
\description{
S3 details for rct objects
}
