% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wk-vctr.R, R/wkb.R
\name{rep_len.wk_vctr}
\alias{rep_len.wk_vctr}
\alias{new_wk_wkb}
\alias{validate_wk_wkb}
\title{S3 Details for wk_wkb}
\usage{
rep_len.wk_vctr(x, ...)

new_wk_wkb(x)

validate_wk_wkb(x)
}
\arguments{
\item{x}{A (possibly) \code{\link[=wkb]{wkb()}} vector}

\item{...}{Unused}
}
\description{
S3 Details for wk_wkb
}
