% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkt.R
\name{new_wk_wkt}
\alias{new_wk_wkt}
\alias{validate_wk_wkt}
\title{S3 Details for wk_wkt}
\usage{
new_wk_wkt(x = character())

validate_wk_wkt(x)
}
\arguments{
\item{x}{A (possibly) \code{\link[=wkt]{wkt()}} vector}
}
\description{
S3 Details for wk_wkt
}
