% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wk-xy.R
\name{new_wk_xy}
\alias{new_wk_xy}
\alias{new_wk_xyz}
\alias{new_wk_xym}
\alias{new_wk_xyzm}
\alias{validate_wk_xy}
\alias{validate_wk_xyz}
\alias{validate_wk_xym}
\alias{validate_wk_xyzm}
\title{S3 details for xy objects}
\usage{
new_wk_xy(x = list(x = double(), y = double()))

new_wk_xyz(x = list(x = double(), y = double(), z = double()))

new_wk_xym(x = list(x = double(), y = double(), m = double()))

new_wk_xyzm(x = list(x = double(), y = double(), z = double(), m = double()))

validate_wk_xy(x)

validate_wk_xyz(x)

validate_wk_xym(x)

validate_wk_xyzm(x)
}
\arguments{
\item{x}{A bare named list.}
}
\value{
A vector of class 'wk_xy(zm)'
}
\description{
S3 details for xy objects
}
