% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct.R
\name{rct}
\alias{rct}
\alias{as_rct}
\alias{as_rct.wk_rct}
\alias{as_rct.matrix}
\alias{as_rct.data.frame}
\title{2D rectangle vectors}
\usage{
rct(
  xmin = double(),
  ymin = double(),
  xmax = double(),
  ymax = double(),
  crs = wk_crs_auto()
)

as_rct(x, ...)

\method{as_rct}{wk_rct}(x, ...)

\method{as_rct}{matrix}(x, ..., crs = NULL)

\method{as_rct}{data.frame}(x, ..., crs = NULL)
}
\arguments{
\item{xmin, ymin, xmax, ymax}{Rectangle bounds.}

\item{crs}{A value to be propagated as the CRS for this vector.}

\item{x}{An object to be converted to a \code{\link[=rct]{rct()}}.}

\item{...}{Extra arguments passed to \code{as_rct()}.}
}
\value{
A vector along the recycled length of bounds.
}
\description{
2D rectangle vectors
}
\examples{
rct(1, 2, 3, 4)

}
