% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct.R
\name{rct_xmin}
\alias{rct_xmin}
\alias{rct_ymin}
\alias{rct_xmax}
\alias{rct_ymax}
\alias{rct_width}
\alias{rct_height}
\alias{rct_intersects}
\alias{rct_contains}
\alias{rct_intersection}
\title{Rectangle accessors and operators}
\usage{
rct_xmin(x)

rct_ymin(x)

rct_xmax(x)

rct_ymax(x)

rct_width(x)

rct_height(x)

rct_intersects(x, y)

rct_contains(x, y)

rct_intersection(x, y)
}
\arguments{
\item{x, y}{\code{\link[=rct]{rct()}} vectors}
}
\value{
\itemize{
\item \code{rct_xmin()}, \code{rct_xmax()}, \code{rct_ymin()}, and \code{rct_ymax()} return
the components of the \code{\link[=rct]{rct()}}.
}
}
\description{
Rectangle accessors and operators
}
\examples{
x <- rct(0, 0, 10, 10)
y <- rct(5, 5, 15, 15)

rct_xmin(x)
rct_ymin(x)
rct_xmax(x)
rct_ymax(x)
rct_height(x)
rct_width(x)
rct_intersects(x, y)
rct_intersection(x, y)
rct_contains(x, y)
rct_contains(x, rct(4, 4, 6, 6))

}
