% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_wkb_wkt}
\alias{translate_wkb_wkt}
\title{Translate between WKB and WKT}
\usage{
translate_wkb_wkt(x)
}
\arguments{
\item{x}{A \code{list()} of \code{\link[=raw]{raw()}} vectors, such as that
returned by \code{\link[sf:st_as_binary]{sf::st_as_binary()}}.}
}
\value{
A character vector of well-known text.
}
\description{
Translate between WKB and WKT
}
