% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-vctrs.R
\name{vctrs-methods}
\alias{vctrs-methods}
\alias{vec_cast.wk_wkb}
\alias{vec_ptype2.wk_wkb}
\alias{vec_cast.wk_wkt}
\alias{vec_ptype2.wk_wkt}
\alias{vec_cast.wk_wksxp}
\alias{vec_ptype2.wk_wksxp}
\alias{vec_cast.wk_xy}
\alias{vec_ptype2.wk_xy}
\alias{vec_cast.wk_xyz}
\alias{vec_ptype2.wk_xyz}
\alias{vec_cast.wk_xym}
\alias{vec_ptype2.wk_xym}
\alias{vec_cast.wk_xyzm}
\alias{vec_ptype2.wk_xyzm}
\title{Vctrs methods}
\usage{
vec_cast.wk_wkb(x, to, ...)

vec_ptype2.wk_wkb(x, y, ...)

vec_cast.wk_wkt(x, to, ...)

vec_ptype2.wk_wkt(x, y, ...)

vec_cast.wk_wksxp(x, to, ...)

vec_ptype2.wk_wksxp(x, y, ...)

vec_cast.wk_xy(x, to, ...)

vec_ptype2.wk_xy(x, y, ...)

vec_cast.wk_xyz(x, to, ...)

vec_ptype2.wk_xyz(x, y, ...)

vec_cast.wk_xym(x, to, ...)

vec_ptype2.wk_xym(x, y, ...)

vec_cast.wk_xyzm(x, to, ...)

vec_ptype2.wk_xyzm(x, y, ...)
}
\arguments{
\item{x, y, to, ...}{See \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and \code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}.}
}
\description{
Vctrs methods
}
