% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wk-package.R
\docType{package}
\name{wk-package}
\alias{wk}
\alias{wk-package}
\title{wk: Lightweight Well-Known Geometry Parsing}
\description{
Provides a minimal R and C++ API for parsing
  well-known binary (WKB) and well-known text (WKT).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/wk}
  \item Report bugs at \url{https://github.com/paleolimbot/wk/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

Authors:
\itemize{
  \item Edzer Pebesma \email{edzer.pebesma@uni-muenster.de} (\href{https://orcid.org/0000-0001-8049-7069}{ORCID})
}

Other contributors:
\itemize{
  \item GEOS developers [copyright holder]
}

}
\keyword{internal}
