% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{wk_bbox}
\alias{wk_bbox}
\alias{wk_bbox.default}
\alias{wk_bbox_handler}
\title{2D bounding rectangles}
\usage{
wk_bbox(handleable)

\method{wk_bbox}{default}(handleable)

wk_bbox_handler()
}
\arguments{
\item{handleable}{A geometry vector (e.g., \code{\link[=wkb]{wkb()}}, \code{\link[=wkt]{wkt()}}, \code{\link[=xy]{xy()}},
\code{\link[=rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[=wk_handle]{wk_handle()}} is defined.}
}
\value{
A \code{\link[=rct]{rct()}} of length 1.
}
\description{
2D bounding rectangles
}
\examples{
wk_bbox(wkt("LINESTRING (1 2, 3 5)"))

}
