% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wk-crs.R
\name{wk_crs}
\alias{wk_crs}
\alias{wk_crs.wk_vctr}
\alias{wk_crs.wk_rcrd}
\alias{wk_crs<-}
\alias{wk_set_crs}
\alias{wk_crs_output}
\alias{wk_geodesic_output}
\title{Set and get vector CRS}
\usage{
wk_crs(x)

\method{wk_crs}{wk_vctr}(x)

\method{wk_crs}{wk_rcrd}(x)

wk_crs(x) <- value

wk_set_crs(x, crs)

wk_crs_output(...)

wk_geodesic_output(...)
}
\arguments{
\item{x, ...}{Objects whose "crs" attribute is used to carry a CRS.}

\item{crs, value}{An object that can be interpreted as a CRS}
}
\description{
The wk package doesn't operate on CRS objects, but does propagate them
through subsetting and concatenation. A CRS object can be any R object,
and x can be any object whose 'crs' attribute carries a CRS. These functions
are S3 generics to keep them from being used
on objects that do not use this system of CRS propagation.
}
