% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wk-crs.R
\name{wk_crs_equal}
\alias{wk_crs_equal}
\alias{wk_crs_output}
\alias{wk_crs_equal_generic}
\alias{wk_crs_inherit}
\title{Compare CRS objects}
\usage{
wk_crs_equal(x, y)

wk_crs_output(x, y)

wk_crs_equal_generic(x, y, ...)

wk_crs_inherit()
}
\arguments{
\item{x, y}{Objects stored in the \code{crs} attribute of a vector.}

\item{...}{Unused}
}
\value{
\code{TRUE} if \code{x} and \code{y} can be considered equal, \code{FALSE} otherwise.
}
\description{
The \code{\link[=wk_crs_equal]{wk_crs_equal()}} function uses special S3 dispatch on \code{\link[=wk_crs_equal_generic]{wk_crs_equal_generic()}}
to evaluate whether or not two CRS values can be considered equal. When implementing
\code{\link[=wk_crs_equal_generic]{wk_crs_equal_generic()}}, every attempt should be made to make \code{wk_crs_equal(x, y)}
and \code{wk_crs_equal(y, x)} return identically.
}
