% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{wk_grid}
\alias{wk_grid}
\alias{wk_grid_rct}
\alias{wk_grid_xy}
\alias{as_wk_grid_rct}
\alias{as_wk_grid_rct.wk_grid_rct}
\alias{as_wk_grid_xy}
\alias{as_wk_grid_xy.wk_grid_xy}
\title{Raster-like objects}
\usage{
wk_grid(
  bbox = NULL,
  nx = NULL,
  ny = NULL,
  dx = NULL,
  dy = NULL,
  type = c("polygons", "corners", "centers")
)

wk_grid_rct(data, bbox = rct(0, 0, dim(data)[1], dim(data)[2]))

wk_grid_xy(data, bbox = rct(0, 0, dim(data)[1], dim(data)[2]))

as_wk_grid_rct(x, ...)

\method{as_wk_grid_rct}{wk_grid_rct}(x, ...)

as_wk_grid_xy(x, ...)

\method{as_wk_grid_xy}{wk_grid_xy}(x, ...)
}
\arguments{
\item{bbox}{A \code{\link[=rct]{rct()}} containing the bounds and CRS of the object.}

\item{nx, ny, dx, dy}{Either a number of cells in the x- and y- directions
or delta in the x- and y-directions (in which case \code{bbox} must
be specified).}

\item{type}{Use "polygons" to return a grid whose objects can be
represented using an \code{\link[=rct]{rct()}}; use "centers" to return a grid whose
objects are the center of the \code{\link[=rct]{rct()}} grid; use "corners" to return
a grid along the corners of \code{bbox}.}

\item{data}{An object with two or more dimensions. Most usefully, a matrix.}

\item{x}{An object to convert to a grid}

\item{...}{Passed to S3 methods}
}
\value{
\itemize{
\item \code{wk_grid()} returns a \code{wk_grid_rct()} for \verb{type == "polygons} or
a \code{wk_grid_xy()} otherwise.
\item \code{wk_grid_rct()} returns an object of class "wk_grid_rct".
\item \code{wk_grid_xy()} returns an object of class "wk_grid_xy".
}
}
\description{
Raster-like objects
}
\examples{
wk_grid_rct(volcano)
# approx bounding box in New Zealand Transverse Mercator
bbox <- wk::rct(5917019, 1756993, 5917466, 1757577, crs = "EPSG:2193")
wk_grid_rct(volcano, bbox)

}
