% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{wk_grid_rct}
\alias{wk_grid_rct}
\alias{wk_grid_xy}
\title{Raster-like objects}
\usage{
wk_grid_rct(data, bbox = rct(0, 0, dim(data)[1], dim(data)[2]))

wk_grid_xy(data, bbox = rct(0, 0, dim(data)[1], dim(data)[2]))
}
\arguments{
\item{data}{An object with two or more dimensions. Most usefully, a matrix.}

\item{bbox}{A \code{\link[=rct]{rct()}} with bounds.}
}
\value{

}
\description{
Raster-like objects
}
\examples{
wk_grid(volcano)
# approx bounding box in New Zealand Transverse Mercator
bbox <- wk::rct(5917019, 1756993, 5917466, 1757577, crs = "EPSG:2193")
wk_grid(volcano, bbox)

}
