% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd-handle.R
\name{wk_handle.grd_xy}
\alias{wk_handle.grd_xy}
\alias{wk_handle.grd_rct}
\title{Handler interface for grid objects}
\usage{
\method{wk_handle}{grd_xy}(handleable, handler, ...)

\method{wk_handle}{grd_rct}(handleable, handler, ...)
}
\arguments{
\item{handleable}{A geometry vector (e.g., \code{\link[=wkb]{wkb()}}, \code{\link[=wkt]{wkt()}}, \code{\link[=xy]{xy()}},
\code{\link[=rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[=wk_handle]{wk_handle()}} is defined.}

\item{handler}{A \link[=wk_handle]{wk_handler} object.}

\item{...}{Passed to the \code{\link[=wk_handle]{wk_handle()}} method.}
}
\value{
The result of the \code{handler}.
}
\description{
Handler interface for grid objects
}
\examples{
wk_handle(grd(nx = 3, ny = 3), wkt_writer())
wk_handle(grd(nx = 3, ny = 3, type = "centers"), wkt_writer())

}
