% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{wk_handle}
\alias{wk_handle}
\alias{wk_handle.wk_wkb}
\alias{wk_handle.wk_wkt}
\alias{wk_handle.wk_xy}
\alias{wk_handle.wk_rct}
\alias{wk_handle.sfc}
\alias{wk_void_handler}
\alias{wk_void}
\alias{new_wk_handler}
\alias{is_wk_handler}
\alias{as_wk_handler}
\title{Read geometry vectors}
\usage{
wk_handle(x, handler, ...)

\method{wk_handle}{wk_wkb}(x, handler, ...)

\method{wk_handle}{wk_wkt}(x, handler, ...)

\method{wk_handle}{wk_xy}(x, handler, ...)

\method{wk_handle}{wk_rct}(x, handler, ...)

\method{wk_handle}{sfc}(x, handler, ...)

wk_void_handler()

wk_void(x, ...)

new_wk_handler(handler_ptr, subclass = character())

is_wk_handler(handler)

as_wk_handler(handler, ...)
}
\arguments{
\item{x}{A geometry vector (e.g., \code{\link[=wkb]{wkb()}} or \code{\link[=wkt]{wkt()}}).}

\item{handler}{An object created with \code{\link[=new_wk_handler]{new_wk_handler()}}.}

\item{...}{Unused}

\item{handler_ptr}{An external pointer to a newly created WK handler}

\item{subclass}{The handler subclass}
}
\value{
A WK handler.
}
\description{
Read geometry vectors
}
