% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wk-crs.R
\name{wk_is_geodesic}
\alias{wk_is_geodesic}
\alias{wk_set_geodesic}
\alias{wk_is_geodesic<-}
\title{Set and get vector geodesic edge interpolation}
\usage{
wk_is_geodesic(x)

wk_set_geodesic(x, geodesic)

wk_is_geodesic(x) <- value
}
\arguments{
\item{x}{An R object that contains edges}

\item{geodesic, value}{\code{TRUE} if edges must be interpolated as geodesics when
coordinates are spherical, \code{FALSE} otherwise.}
}
\value{
\code{TRUE} if edges must be interpolated as geodesics when
coordinates are spherical, \code{FALSE} otherwise.
}
\description{
Set and get vector geodesic edge interpolation
}
