% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{wk_make_linestring}
\alias{wk_make_linestring}
\alias{wk_make_polygon}
\alias{wk_make_collection}
\alias{wk_linestring_filter}
\alias{wk_polygon_filter}
\alias{wk_collection_filter}
\title{Create lines, polygons, and collections}
\usage{
wk_make_linestring(handleable, feature_id = 1L, ...)

wk_make_polygon(handleable, feature_id = 1L, ring_id = 1L, ...)

wk_make_collection(handleable, geometry_type = 7L, feature_id = 1L, ...)

wk_linestring_filter(handler, feature_id = 1L)

wk_polygon_filter(handler, feature_id = 1L, ring_id = 1L)

wk_collection_filter(handler, feature_id = 1L, geometry_type = 7L)
}
\arguments{
\item{handleable}{A geometry vector (e.g., \code{\link[=wkb]{wkb()}}, \code{\link[=wkt]{wkt()}}, \code{\link[=xy]{xy()}},
\code{\link[=rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[=wk_handle]{wk_handle()}} is defined.}

\item{feature_id}{An identifier where changes in sequential
values indicate a new feature.}

\item{...}{Passed to the \code{\link[=wk_handle]{wk_handle()}} method.}

\item{ring_id}{An identifier where changes in sequential
values indicate a new ring. Rings are automatically
closed.}

\item{geometry_type}{The collection type to create.}

\item{handler}{A \link[=wk_handle]{wk_handler} object.}
}
\value{
An object of the same class as \code{handleable} with
whose coordinates have been assembled into the given
type.
}
\description{
Create lines, polygons, and collections
}
\examples{
wk_make_linestring(xy(c(1, 1), c(2, 3)))
wk_make_polygon(xy(c(0, 1, 1), c(0, 0, 1)))
wk_make_collection(xy(c(1, 1), c(2, 3)))

}
