% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{wk_meta}
\alias{wk_meta}
\alias{wk_meta.default}
\alias{wk_meta_handler}
\title{Extract feature-level meta}
\usage{
wk_meta(handleable, ...)

\method{wk_meta}{default}(handleable, ...)

wk_meta_handler()
}
\arguments{
\item{handleable}{A geometry vector (e.g., \code{\link[=wkb]{wkb()}}, \code{\link[=wkt]{wkt()}}, \code{\link[=xy]{xy()}},
\code{\link[=rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[=wk_handle]{wk_handle()}} is defined.}

\item{...}{Passed to the \code{\link[=wk_handle]{wk_handle()}} method.}
}
\value{
A data.frame with columns:
\itemize{
\item \code{geometry_type}: An integer identifying the geometry type
\item \code{size}: For points and linestrings, the number of coordinates; for
polygons, the number of rings; for collections, the number of
child geometries. A value of zero indicates an EMPTY geometry.
A value of \code{NA} means this value is unknown without parsing the
entire geometry.
\item \code{has_z}: \code{TRUE} if coordinates contain a Z value
\item \code{has_m}: \code{TRUE} if coordinates contain an M value
\item \code{srid}: An integer identifying a CRS or NA if this value was not
provided.
\item \code{precision}: A grid size or 0.0 if a grid size was not provided.
}
}
\description{
Extract feature-level meta
}
\examples{
wk_meta(as_wkt("LINESTRING (0 0, 1 1)"))
wk_meta(as_wkb("LINESTRING (0 0, 1 1)"))

}
