% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wk_proj_crs_view}
\alias{wk_proj_crs_view}
\alias{wk_proj_crs_json}
\title{Common CRS Representations}
\format{
An object of class \code{data.frame} with 13096 rows and 7 columns.

An object of class \code{data.frame} with 13096 rows and 3 columns.
}
\usage{
wk_proj_crs_view

wk_proj_crs_json
}
\description{
These fixtures are calculated from PROJ version 9.1.0 and the database
built from its source. They are used internally to transform and inspect
coordinate reference systems.
}
\examples{
head(wk_proj_crs_view)
colnames(wk_proj_crs_json)

}
\keyword{datasets}
