% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-data-frame.R, R/slice.R
\name{wk_slice.data.frame}
\alias{wk_slice.data.frame}
\alias{wk_slice}
\alias{wk_slice.default}
\title{Handle specific regions of objects}
\usage{
\method{wk_slice}{data.frame}(
  handleable,
  from = NULL,
  to = NULL,
  times = 1L,
  each = 1L,
  handler = wk_writer(handleable),
  ...
)

wk_slice(
  handleable,
  from = NULL,
  to = NULL,
  times = 1L,
  each = 1L,
  handler = wk_writer(handleable),
  ...
)

\method{wk_slice}{default}(
  handleable,
  from = NULL,
  to = NULL,
  times = 1L,
  each = 1L,
  handler = wk_writer(handleable),
  ...
)
}
\arguments{
\item{handleable}{A geometry vector (e.g., \code{\link[=wkb]{wkb()}}, \code{\link[=wkt]{wkt()}}, \code{\link[=xy]{xy()}},
\code{\link[=rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[=wk_handle]{wk_handle()}} is defined.}

\item{from}{1-based index of the feature to start from}

\item{to}{1-based index of the feature to end at}

\item{times}{The number of times the entire handleable should be
repeated.}

\item{each}{The number of times each feature should be repeated.}

\item{handler}{A \link[=wk_handle]{wk_handler} object.}

\item{...}{Passed to the \code{\link[=wk_handle]{wk_handle()}} method.}
}
\value{
A subset of \code{handleable}
}
\description{
Handle specific regions of objects
}
\examples{
wk_slice(xy(1:5, 1:5), from = 3, to = 5)
wk_slice(
  data.frame(let = letters[1:5], geom = xy(1:5, 1:5)),
  from = 3, to = 5
)

}
