% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{wk_trans_inverse}
\alias{wk_trans_inverse}
\title{Generic transform class}
\usage{
wk_trans_inverse(trans, ...)
}
\arguments{
\item{trans}{An external pointer to a wk_trans object}

\item{...}{Passed to S3 methods}

\item{trans_ptr}{An external pointer to a wk_trans_t transform
struct.}

\item{subclass}{An optional subclass to apply to the pointer}
}
\description{
Generic transform class
}
