% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-sf.R, R/translate.R
\name{wk_translate.sfc}
\alias{wk_translate.sfc}
\alias{wk_translate}
\alias{wk_translate.default}
\title{Translate geometry vectors}
\usage{
\method{wk_translate}{sfc}(handleable, to, ...)

wk_translate(handleable, to, ...)

\method{wk_translate}{default}(handleable, to, ...)
}
\arguments{
\item{handleable}{A geometry vector (e.g., \code{\link[=wkb]{wkb()}}, \code{\link[=wkt]{wkt()}}, \code{\link[=xy]{xy()}},
\code{\link[=rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[=wk_handle]{wk_handle()}} is defined.}

\item{to}{A prototype object.}

\item{...}{Passed to the \code{\link[=wk_handle]{wk_handle()}} method.}
}
\description{
Translate geometry vectors
}
