% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{wk_translate_wkb_wkt}
\alias{wk_translate_wkb_wkt}
\alias{wk_translate_wkb_wkb}
\alias{wk_platform_endian}
\title{Translate between WKB and WKT}
\usage{
wk_translate_wkb_wkt(wkb, precision = 16, trim = TRUE)

wk_translate_wkb_wkb(wkb, endian = wk_platform_endian(), buffer_size = 2048)

wk_platform_endian()
}
\arguments{
\item{wkb}{A \code{list()} of \code{\link[=raw]{raw()}} vectors, such as that
returned by \code{\link[sf:st_as_binary]{sf::st_as_binary()}}.}

\item{precision}{The rounding precision to use when writing
(number of decimal places).}

\item{trim}{Trim unnecessary zeroes in the output?}

\item{endian}{For WKB writing, 0 for big endian, 1 for little endinan.
Defaults to \code{\link[=wk_platform_endian]{wk_platform_endian()}}.}

\item{buffer_size}{For WKB writing, the initial buffer size to use for
each feature, in bytes. This will be extended when needed, but if you
are calling this repeatedly with huge geometries, setting this value
to a larger number may result in less copying.}
}
\value{
\code{\link[=wk_translate_wkb_wkt]{wk_translate_wkb_wkt()}} returns a character vector of
well-known text; \code{\link[=wk_translate_wkb_wkb]{wk_translate_wkb_wkb()}} returns a list
of raw vectors.
}
\description{
Translate between WKB and WKT
}
