% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex-filter.R
\name{wk_vertices}
\alias{wk_vertices}
\alias{wk_vertex_filter}
\title{Extract vertices}
\usage{
wk_vertices(handleable, ...)

wk_vertex_filter(handler)
}
\arguments{
\item{handleable}{A geometry vector (e.g., \code{\link[=wkb]{wkb()}}, \code{\link[=wkt]{wkt()}}, \code{\link[=xy]{xy()}},
\code{\link[=rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[=wk_handle]{wk_handle()}} is defined.}

\item{...}{Passed to the \code{\link[=wk_handle]{wk_handle()}} method.}

\item{handler}{A \link[=wk_handle]{wk_handler} object.}
}
\value{
The result of the \code{handler}.
}
\description{
Extract vertices
}
\examples{
wk_vertices(wkt("LINESTRING (0 0, 1 1)"))

}
