% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/void.R
\name{wk_void}
\alias{wk_void}
\alias{wk_void_handler}
\title{Do nothing}
\usage{
wk_void(handleable, ...)

wk_void_handler()
}
\arguments{
\item{handleable}{A geometry vector (e.g., \code{\link[=wkb]{wkb()}}, \code{\link[=wkt]{wkt()}}, \code{\link[=xy]{xy()}},
\code{\link[=rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[=wk_handle]{wk_handle()}} is defined.}

\item{...}{Passed to the \code{\link[=wk_handle]{wk_handle()}} method.}
}
\value{
\code{NULL}
}
\description{
This handler does nothing and returns \code{NULL}. It is useful for
benchmarking readers and handlers and when using filters
that have side-effects (e.g., \code{\link[=wk_debug]{wk_debug()}}). Note that this
handler stops on the first parse error; to see a list of parse
errors see the \code{\link[=wk_problems]{wk_problems()}} handler.
}
\examples{
wk_void(wkt("POINT (1 4)"))
wk_handle(wkt("POINT (1 4)"), wk_void_handler())

}
