% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wk_writer}
\alias{wk_writer}
\alias{wk_writer.wk_wkt}
\alias{wk_writer.wk_wkb}
\alias{wk_writer.sfc}
\alias{wk_writer.wk_xy}
\alias{wkt_writer}
\alias{wkb_writer}
\alias{sfc_writer}
\alias{xyzm_writer}
\title{Write geometry vectors}
\usage{
wk_writer(handleable, ...)

\method{wk_writer}{wk_wkt}(handleable, ..., precision = 16, trim = TRUE)

\method{wk_writer}{wk_wkb}(handleable, ...)

\method{wk_writer}{sfc}(handleable, ...)

\method{wk_writer}{wk_xy}(handleable, ...)

wkt_writer(precision = 16, trim = TRUE)

wkb_writer()

sfc_writer()

xyzm_writer()
}
\arguments{
\item{handleable}{Placeholder}

\item{...}{For S3 generic compatibility}

\item{precision}{If \code{trim} is \code{TRUE}, the total number of significant digits to keep
for each result or the number of digits after the decimal place otherwise.}

\item{trim}{Use \code{FALSE} to keep trailing zeroes after the decimal place.}
}
\value{
A \link[=wk_handle]{wk_handler}.
}
\description{
Write geometry vectors
}
