% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{wkb_coords}
\alias{wkb_coords}
\alias{wkt_coords}
\title{Data frames of coordinates}
\usage{
wkb_coords(wkb)

wkt_coords(wkt)
}
\arguments{
\item{wkb}{A \code{list()} of \code{\link[=raw]{raw()}} vectors, such as that
returned by \code{\link[sf:st_as_binary]{sf::st_as_binary()}}.}

\item{wkt}{A character vector containing well-known text.}
}
\value{
A data.frame with columns:
\itemize{
\item \code{feature_id}
\item \code{nest_id}
\item \code{part_id}
\item \code{ring_id}
\item \code{coordinate_id}
\item \code{x}
\item \code{y}
\item \code{z}
\item \code{m}
}
}
\description{
Data frames of coordinates
}
