% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{wkb_format}
\alias{wkb_format}
\alias{wkt_format}
\title{Deprecated functions}
\usage{
wkb_format(wkb, max_coords = 3, precision = 6, trim = TRUE)

wkt_format(wkt, max_coords = 3, precision = 6, trim = TRUE)
}
\arguments{
\item{wkb}{A \code{list()} of \code{\link[=raw]{raw()}} vectors, such as that
returned by \code{sf::st_as_binary()}.}

\item{max_coords}{The maximum number of coordinates to include
in the output.}

\item{precision}{The rounding precision to use when writing
(number of decimal places).}

\item{trim}{Trim unnecessary zeroes in the output?}

\item{wkt}{A character vector containing well-known text.}
}
\description{
These functions are deprecated and will be removed in a future version.
}
