% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{wkb_problems}
\alias{wkb_problems}
\title{Validate well-known binary}
\usage{
wkb_problems(wkb)
}
\arguments{
\item{wkb}{A \code{list()} of \code{\link[=raw]{raw()}} vectors, such as that
returned by \code{\link[sf:st_as_binary]{sf::st_as_binary()}}.}
}
\value{
A character vector of parsing errors. \code{NA} signifies
that there was no parsing error.
}
\description{
Validate well-known binary
}
