% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkb.R
\name{wkb_to_hex}
\alias{wkb_to_hex}
\title{Convert well-known binary to hex}
\usage{
wkb_to_hex(x)
}
\arguments{
\item{x}{A \code{\link[=wkb]{wkb()}} vector}
}
\value{
A hex encoded \code{\link[=wkb]{wkb()}} vector
}
\description{
Convert well-known binary to hex
}
\examples{
x <- as_wkb(xyz(1:5, 6:10, 11:15))
wkb_to_hex(x)

}
