% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{wkb_translate_wkt}
\alias{wkb_translate_wkt}
\alias{wkb_translate_wkb}
\alias{wkt_translate_wkt}
\alias{wkt_translate_wkb}
\alias{wk_platform_endian}
\title{Translate between WKB and WKT}
\usage{
wkb_translate_wkt(wkb, ..., precision = 16, trim = TRUE)

wkb_translate_wkb(wkb, ...)

wkt_translate_wkt(wkt, ..., precision = 16, trim = TRUE)

wkt_translate_wkb(wkt, ...)

wk_platform_endian()
}
\arguments{
\item{wkb}{A \code{list()} of \code{\link[=raw]{raw()}} vectors, such as that
returned by \code{sf::st_as_binary()}.}

\item{...}{Used to keep backward compatibility with previous
versions of these functions.}

\item{precision}{The rounding precision to use when writing
(number of decimal places).}

\item{trim}{Trim unnecessary zeroes in the output?}

\item{wkt}{A character vector containing well-known text.}
}
\value{
\verb{*_translate_wkt()} returns a character vector of
well-known text; \verb{*_translate_wkb()} returns a list
of raw vectors. Unlike \code{\link[=as_wkb]{as_wkb()}}, \code{\link[=as_wkt]{as_wkt()}},
these functions do not attach a class to the output.
}
\description{
Translate between WKB and WKT
}
\examples{
# translate between WKT and WKB
(wkb <- wkt_translate_wkb("POINT (30 10)"))
wkb_translate_wkt(wkb)

# some basic creation options are also available
wkt_translate_wkt("POINT (30 10)", trim = FALSE)

}
