% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkbheaders-package.R
\docType{package}
\name{wkbheaders-package}
\alias{wkbheaders}
\alias{wkbheaders-package}
\title{wkbheaders: C++ Headers for Well-Known Binary Conversion}
\description{
Provides a small header-only library for reading and writing
  well-known binary (WKB) to and from any R-based format.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
