% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkheaders-package.R
\docType{package}
\name{wkheaders-package}
\alias{wkheaders}
\alias{wkheaders-package}
\title{wkheaders: C++ Headers for Well-Known Geometry Conversion}
\description{
Provides a small header-only library for reading and writing
  well-known binary (WKB) and well-known text (WKT) to and from any R-based 
  format.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
