% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wk-xy.R
\name{xy}
\alias{xy}
\alias{xyz}
\alias{xym}
\alias{xyzm}
\alias{xy_dims}
\alias{as_xy}
\alias{as_xy.wk_xy}
\title{Efficient point vectors}
\usage{
xy(x = double(), y = double())

xyz(x = double(), y = double(), z = double())

xym(x = double(), y = double(), m = double())

xyzm(x = double(), y = double(), z = double(), m = double())

xy_dims(x)

as_xy(x, ...)

\method{as_xy}{wk_xy}(x, ..., dims = NULL)
}
\arguments{
\item{x, y, z, m}{Coordinate values.}

\item{...}{Passed to methods.}

\item{dims}{A set containing one or more of \code{c("x", "y", "z", "m")}.}
}
\value{
A vector of coordinate values.
}
\description{
Efficient point vectors
}
\examples{
xy(1:5, 1:5)
xyz(1:5, 1:5, 10)
xym(1:5, 1:5, 10)
xyzm(1:5, 1:5, 10, 12)

}
