% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyzm.R
\name{xy_x}
\alias{xy_x}
\alias{xy_y}
\alias{xy_z}
\alias{xy_m}
\title{XY vector extractors}
\usage{
xy_x(x)

xy_y(x)

xy_z(x)

xy_m(x)
}
\arguments{
\item{x}{An \code{\link[=xy]{xy()}} vector}
}
\value{
Components of the \code{\link[=xy]{xy()}} vector or NULL if the dimension is missing
}
\description{
XY vector extractors
}
\examples{
x <- xyz(1:5, 6:10, 11:15)
xy_x(x)
xy_y(x)
xy_z(x)
xy_m(x)

}
