% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{wkb_translate_wkt}
\alias{wkb_translate_wkt}
\alias{wkb_translate_wkb}
\alias{wkt_translate_wkt}
\alias{wkt_translate_wkb}
\title{Deprecated functions}
\usage{
wkb_translate_wkt(wkb, ..., precision = 16, trim = TRUE)

wkb_translate_wkb(wkb, ..., endian = NA_integer_)

wkt_translate_wkt(wkt, ..., precision = 16, trim = TRUE)

wkt_translate_wkb(wkt, ..., endian = NA_integer_)
}
\arguments{
\item{wkb}{A \code{list()} of \code{\link[=raw]{raw()}} vectors, such as that
returned by \code{sf::st_as_binary()}.}

\item{...}{Used to keep backward compatibility with previous
versions of these functions.}

\item{precision}{The rounding precision to use when writing
(number of decimal places).}

\item{trim}{Trim unnecessary zeroes in the output?}

\item{endian}{Force the endian of the resulting WKB.}

\item{wkt}{A character vector containing well-known text.}

\item{max_coords}{The maximum number of coordinates to include
in the output.}
}
\description{
These functions are deprecated and will be removed in a future version.
}
