% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd.R
\name{grd}
\alias{grd}
\alias{grd_rct}
\alias{grd_xy}
\alias{as_grd_rct}
\alias{as_grd_rct.wk_grd_rct}
\alias{grd_subset}
\alias{as_grd_rct.wk_grd_xy}
\alias{as_grd_xy}
\alias{as_grd_xy.wk_grd_xy}
\alias{as_grd_xy.wk_grd_rct}
\alias{grd_data_spec}
\alias{grd_data_spec.default}
\alias{grd_data_spec.nativeRaster}
\title{Raster-like objects}
\usage{
grd(
  bbox = NULL,
  nx = NULL,
  ny = NULL,
  dx = NULL,
  dy = NULL,
  type = c("polygons", "corners", "centers")
)

grd_rct(
  data,
  bbox = rct(0, 0, dim(data)[2], dim(data)[1]),
  spec = grd_data_spec(data)
)

grd_xy(
  data,
  bbox = rct(0, 0, dim(data)[2] - 1, dim(data)[1] - 1),
  spec = grd_data_spec(data)
)

as_grd_rct(x, ...)

\method{as_grd_rct}{wk_grd_rct}(x, ...)

grd_subset(object, x = NULL, y = NULL, ..., bbox = NULL)

\method{as_grd_rct}{wk_grd_xy}(x, ...)

as_grd_xy(x, ...)

\method{as_grd_xy}{wk_grd_xy}(x, ...)

\method{as_grd_xy}{wk_grd_rct}(x, ...)

grd_data_spec(data, ...)

\method{grd_data_spec}{default}(data, ...)

\method{grd_data_spec}{nativeRaster}(data, ...)
}
\arguments{
\item{bbox}{A \code{\link[=rct]{rct()}} containing the bounds and CRS of the object.}

\item{nx, ny, dx, dy}{Either a number of cells in the x- and y- directions
or delta in the x- and y-directions (in which case \code{bbox} must
be specified).}

\item{type}{Use "polygons" to return a grid whose objects can be
represented using an \code{\link[=rct]{rct()}}; use "centers" to return a grid whose
objects are the center of the \code{\link[=rct]{rct()}} grid; use "corners" to return
a grid along the corners of \code{bbox}.}

\item{data}{An object with two or more dimensions. Most usefully, a matrix.}

\item{spec}{A \code{\link[=grd_data_spec]{grd_data_spec()}}, used if \code{data} has a non-standard dimension
ordering or subset method implementation.}

\item{x}{An object to convert to a grid}

\item{...}{Passed to S3 methods}
}
\value{
\itemize{
\item \code{grd()} returns a \code{grd_rct()} for \verb{type == "polygons} or
a \code{grd_xy()} otherwise.
\item \code{grd_rct()} returns an object of class "grd_rct".
\item \code{grd_xy()} returns an object of class "grd_xy".
}
}
\description{
Raster-like objects
}
\examples{
grd_rct(volcano)
# approx bounding box in New Zealand Transverse Mercator
bbox <- rct(
  5917000,       1757000 + 870,
  5917000 + 610, 1757000,
  crs = "EPSG:2193"
)
grd_rct(volcano, bbox)

}
