% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd.R
\name{grd_subset}
\alias{grd_subset}
\title{Subset grd objects}
\usage{
grd_subset(object, x = NULL, y = NULL, ..., bbox = NULL)
}
\arguments{
\item{object}{A \code{\link[=grd]{grd()}}}

\item{x}{Indices in the x direction}

\item{y}{Indices in the y direction}

\item{...}{Pased to the subset method}

\item{bbox}{A bounding box to use as a subset}
}
\value{
A modified \code{\link[=grd]{grd()}}
}
\description{
Subset grd objects
}
\examples{
grid <- grd_rct(volcano)
grd_subset(grid, seq(2, 61, by = 4), seq(2, 87, by = 4))

}
