% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd-subset.R
\name{grd_subset}
\alias{grd_subset}
\alias{grd_subset_indices}
\title{Subset grd objects}
\usage{
grd_subset(object, i = NULL, j = NULL, bbox = NULL, ...)

grd_subset_indices(object, i = NULL, j = NULL, bbox = NULL, ...)
}
\arguments{
\item{object}{A \code{\link[=grd]{grd()}}}

\item{i, j}{Raw indices. These must be equally
spaced if passed as numeric; if passed as logical they are
recycled silently along each dimension.}

\item{bbox}{A bounding box to use as a subset. This is used
to calculate a suitable \code{y} and \code{x} index vector representing
all cells that intersect the \code{bbox}. Cells that only touch \code{bbox}
on the bottom and right are not included in the subset, meaning you
can safely tile a regularly-spaced grid along \code{object} without
double-counting cells.}

\item{...}{Passed to subset methods}
}
\value{
A modified \code{\link[=grd]{grd()}}.
}
\description{
The \code{\link[=grd_subset]{grd_subset()}} method handles the subsetting of a \code{\link[=grd]{grd()}}
in x-y space. Ordering of indices is not considered and logical
indies are recycled silently along dimensions. The result of
a \code{\link[=grd_subset]{grd_subset()}} is always a \code{\link[=grd]{grd()}} of the same type whose
relationship to x-y space has not changed.
}
\examples{
grid <- grd_rct(volcano)
grd_subset(grid, seq(2, 61, by = 4), seq(2, 87, by = 4))

}
