% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd.R
\name{grd_summary}
\alias{grd_summary}
\title{Grid information}
\usage{
grd_summary(grid)
}
\arguments{
\item{grid}{A \code{\link[=grd_xy]{grd_xy()}}, \code{\link[=grd_rct]{grd_rct()}}, or other object
implementing \verb{grd_*()} methods.}
}
\value{
\itemize{
\item \code{grd_summary()} returns a \code{list()} with components
\code{xmin}, \code{ymin}, \code{xmax}, \code{ymax},
\code{nx}, \code{ny}, \code{dx}, \code{dy}, \code{width}, and \code{height}.
}
}
\description{
Grid information
}
\examples{
grd_summary(grd(nx = 3, ny = 2))

}
