% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd.R
\name{new_wk_grd}
\alias{new_wk_grd}
\title{S3 details for grid objects}
\usage{
new_wk_grd(x, subclass = character())
}
\arguments{
\item{x}{A \code{\link[=grd]{grd()}}}

\item{subclass}{An optional subclass.}
}
\value{
An object inheriting from 'grd'
}
\description{
S3 details for grid objects
}
