% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd.R
\name{new_wk_grd_data_spec}
\alias{new_wk_grd_data_spec}
\title{Grid detail specifications}
\usage{
new_wk_grd_data_spec(
  dims,
  bbox = NULL,
  index_order = c("y", "x"),
  dim_order = index_order,
  data_order = dim_order
)
}
\arguments{
\item{dims}{The dimensions of the raster}

\item{bbox}{A \code{\link[=rct]{rct()}} containing the bounds and CRS of the object.}

\item{index_order}{Mapping of dimensions to index axes.
Default ("y", "x") reflects that of \code{\link[grDevices:as.raster]{grDevices::as.raster()}}
and \code{\link[graphics:rasterImage]{graphics::rasterImage()}}.}

\item{dim_order}{Mapping of dimensions to}
}
\value{
An object of class "wk_grd_data_spec"
}
\description{
Wraps the details of the \code{data[]}, \code{dim(data)}, to allow multiple types
of data to back a \code{\link[=grd]{grd()}}. Notably, this allows a nativeRaster (whose
data is defined in row-major order but dimensions report column-major
order) to back a \code{\link[=grd]{grd()}}.
}
\examples{
grd_data_spec(matrix())

}
