% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wk-rct.R
\name{new_wk_rct}
\alias{new_wk_rct}
\title{S3 details for rct objects}
\usage{
new_wk_rct(x, crs = NULL)
}
\arguments{
\item{x}{A \code{\link[=rct]{rct()}}}

\item{crs}{A value to be propagated as the CRS for this vector.}
}
\description{
S3 details for rct objects
}
