% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkb.R
\name{new_wk_wkb}
\alias{new_wk_wkb}
\alias{validate_wk_wkb}
\title{S3 Details for wk_wkb}
\usage{
new_wk_wkb(x = list())

validate_wk_wkb(x)
}
\arguments{
\item{x}{A (possibly) \code{\link[=wkb]{wkb()}} vector}
}
\description{
S3 Details for wk_wkb
}
