% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wksxp.R
\name{new_wk_wksxp}
\alias{new_wk_wksxp}
\alias{validate_wk_wksxp}
\title{S3 Details for wk_wksxp}
\usage{
new_wk_wksxp(x)

validate_wk_wksxp(x)
}
\arguments{
\item{x}{A (possibly) \code{\link[=wksxp]{wksxp()}} vector}
}
\description{
S3 Details for wk_wksxp
}
