% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkt.R
\name{new_wk_wkt}
\alias{new_wk_wkt}
\alias{is_wk_wkt}
\alias{validate_wk_wkt}
\title{S3 Details for wk_wkt}
\usage{
new_wk_wkt(x = character(), crs = NULL, is_geodesic = NULL)

is_wk_wkt(x)

validate_wk_wkt(x)
}
\arguments{
\item{x}{A (possibly) \code{\link[=wkt]{wkt()}} vector}

\item{crs}{A value to be propagated as the CRS for this vector.}

\item{is_geodesic}{\code{TRUE} if edges must be interpolated as geodesics when
coordinates are spherical, \code{FALSE} otherwise.}
}
\description{
S3 Details for wk_wkt
}
