% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyzm.R
\name{new_wk_xy}
\alias{new_wk_xy}
\alias{new_wk_xyz}
\alias{new_wk_xym}
\alias{new_wk_xyzm}
\alias{validate_wk_xy}
\alias{validate_wk_xyz}
\alias{validate_wk_xym}
\alias{validate_wk_xyzm}
\title{S3 details for xy objects}
\usage{
new_wk_xy(x = list(x = double(), y = double()), crs = NULL)

new_wk_xyz(x = list(x = double(), y = double(), z = double()), crs = NULL)

new_wk_xym(x = list(x = double(), y = double(), m = double()), crs = NULL)

new_wk_xyzm(
  x = list(x = double(), y = double(), z = double(), m = double()),
  crs = NULL
)

validate_wk_xy(x)

validate_wk_xyz(x)

validate_wk_xym(x)

validate_wk_xyzm(x)
}
\arguments{
\item{x}{A \code{\link[=xy]{xy()}} object.}

\item{crs}{A value to be propagated as the CRS for this vector.}
}
\description{
S3 details for xy objects
}
