% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wk-rct.R
\name{rct}
\alias{rct}
\alias{as_rct}
\alias{as_rct.wk_rct}
\alias{as_rct.matrix}
\alias{as_rct.data.frame}
\title{2D rectangle vectors}
\usage{
rct(xmin = double(), ymin = double(), xmax = double(), ymax = double())

as_rct(x, ...)

\method{as_rct}{wk_rct}(x, ...)

\method{as_rct}{matrix}(x, ...)

\method{as_rct}{data.frame}(x, ...)
}
\arguments{
\item{xmin, ymin, xmax, ymax}{Rectangle bounds.}

\item{x}{An object to be converted to a \code{\link[=rct]{rct()}}.}

\item{...}{Extra arguments passed to \code{as_rct()}.}
}
\value{
A vector along the recycled length of bounds.
}
\description{
2D rectangle vectors
}
\examples{
rct(1, 2, 3, 4)

}
