% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-vctrs.R
\name{vctrs-methods}
\alias{vctrs-methods}
\alias{vec_cast.wk_wkb}
\alias{vec_ptype2.wk_wkb}
\alias{vec_cast.wk_wkt}
\alias{vec_ptype2.wk_wkt}
\alias{vec_cast.wk_wksxp}
\alias{vec_ptype2.wk_wksxp}
\title{Vctrs methods}
\usage{
vec_cast.wk_wkb(x, to, ...)

vec_ptype2.wk_wkb(x, y, ...)

vec_cast.wk_wkt(x, to, ...)

vec_ptype2.wk_wkt(x, y, ...)

vec_cast.wk_wksxp(x, to, ...)

vec_ptype2.wk_wksxp(x, y, ...)
}
\arguments{
\item{x, y, to, ...}{See \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and \code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}.}
}
\description{
Vctrs methods
}
