% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wk-package.R
\docType{package}
\name{wk-package}
\alias{wk}
\alias{wk-package}
\title{wk: Lightweight Well-Known Geometry Parsing}
\description{
Provides a minimal R and C++ API for parsing
  well-known binary and well-known text representation of
  geometries to and from R-native formats. 
  Well-known binary is compact
  and fast to parse; well-known text is human-readable
  and is useful for writing tests. These formats are only
  useful in R if the information they contain can be 
  accessed in R, for which high-performance functions 
  are provided here.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paleolimbot.github.io/wk/}
  \item \url{https://github.com/paleolimbot/wk}
  \item Report bugs at \url{https://github.com/paleolimbot/wk/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

Authors:
\itemize{
  \item Edzer Pebesma \email{edzer.pebesma@uni-muenster.de} (\href{https://orcid.org/0000-0001-8049-7069}{ORCID})
}

}
\keyword{internal}
