% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{wk_debug}
\alias{wk_debug}
\alias{wk_debug_filter}
\title{Debug filters and handlers}
\usage{
wk_debug(handleable, handler = wk_void_handler())

wk_debug_filter(handler = wk_void_handler())
}
\arguments{
\item{handleable}{Any object that can be iterated over by \code{\link[=wk_handle]{wk_handle()}}.}

\item{handler}{A \link[=wk_handle]{wk_handler}.}
}
\value{
The result of the \code{handler}.
}
\description{
Debug filters and handlers
}
\examples{
wk_debug(wkt("POINT (1 1)"))
wk_handle(wkt("POINT (1 1)"), wk_debug_filter())

}
