% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{wk_format}
\alias{wk_format}
\alias{wkt_format_handler}
\title{Format well-known geometry for printing}
\usage{
wk_format(handleable, precision = 7, trim = TRUE, max_coords = 6)

wkt_format_handler(precision = 7, trim = TRUE, max_coords = 6)
}
\arguments{
\item{handleable}{Any object that can be iterated over by \code{\link[=wk_handle]{wk_handle()}}.}

\item{precision}{The rounding precision to use when writing
(number of decimal places).}

\item{trim}{Trim unnecessary zeroes in the output?}

\item{max_coords}{The maximum number of coordinates to include
in the output.}
}
\value{
A character vector of abbreviated well-known text.
}
\description{
Provides an abbreviated version of the well-known text
representation of a geometry. This returns a constant
number of coordinates for each geometry, so is safe to
use for geometry vectors with many (potentially large)
features.
}
\examples{
wk_format(wkt("MULTIPOLYGON (((0 0, 10 0, 0 10, 0 0)))"))
wk_handle(
  wkt("MULTIPOLYGON (((0 0, 10 0, 0 10, 0 0)))"),
  wkt_format_handler()
)

}
