% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-data-frame.R
\name{wk_handle.data.frame}
\alias{wk_handle.data.frame}
\alias{wk_translate.data.frame}
\title{Use data.frame with wk}
\usage{
\method{wk_handle}{data.frame}(handleable, handler, ..., .env = parent.frame())

\method{wk_translate}{data.frame}(handleable, to, ..., .env = parent.frame())
}
\arguments{
\item{handleable}{A geometry vector (e.g., \code{\link[=wkb]{wkb()}}, \code{\link[=wkt]{wkt()}}, \code{\link[=xy]{xy()}},
\code{\link[=rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[=wk_handle]{wk_handle()}} is defined.}

\item{handler}{A \link[=wk_handle]{wk_handler} object.}

\item{...}{Passed to the \code{\link[=wk_handle]{wk_handle()}} method.}

\item{.env}{Passed to \code{\link[=getS3method]{getS3method()}}, which is used to find
the column in a \code{\link[=data.frame]{data.frame()}} for which a \code{\link[=wk_handle]{wk_handle()}}
method is defined.}

\item{to}{A prototype object.}
}
\description{
Use data.frame with wk
}
\examples{
wk_handle(data.frame(a = wkt("POINT (0 1)")), wkb_writer())
wk_translate(wkt("POINT (0 1)"), data.frame(col_name = wkb()))
wk_translate(data.frame(a = wkt("POINT (0 1)")), data.frame(wkb()))

}
