% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{wk_handle}
\alias{wk_handle}
\alias{wk_handle.wk_wkb}
\alias{wk_handle.wk_wkt}
\alias{wk_handle.wk_xy}
\alias{wk_handle.wk_rct}
\alias{wk_handle.sfc}
\alias{new_wk_handler}
\alias{is_wk_handler}
\alias{as_wk_handler}
\title{Read geometry vectors}
\usage{
wk_handle(handleable, handler, ...)

\method{wk_handle}{wk_wkb}(handleable, handler, ...)

\method{wk_handle}{wk_wkt}(handleable, handler, ...)

\method{wk_handle}{wk_xy}(handleable, handler, ...)

\method{wk_handle}{wk_rct}(handleable, handler, ...)

\method{wk_handle}{sfc}(handleable, handler, ...)

new_wk_handler(handler_ptr, subclass = character())

is_wk_handler(handler)

as_wk_handler(handler, ...)
}
\arguments{
\item{handleable}{A geometry vector (e.g., \code{\link[=wkb]{wkb()}}, \code{\link[=wkt]{wkt()}}, \code{\link[=xy]{xy()}},
\code{\link[=rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[=wk_handle]{wk_handle()}} is defined.}

\item{handler}{A \link[=wk_handle]{wk_handler} object.}

\item{...}{Passed to the \code{\link[=wk_handle]{wk_handle()}} method.}

\item{handler_ptr}{An external pointer to a newly created WK handler}

\item{subclass}{The handler subclass}
}
\value{
A WK handler.
}
\description{
Read geometry vectors
}
