% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{wk_meta}
\alias{wk_meta}
\alias{wk_meta_handler}
\alias{wk_vector_meta}
\alias{wk_vector_meta_handler}
\title{Extract vector- and feature-level meta}
\usage{
wk_meta(x)

wk_meta_handler()

wk_vector_meta(x)

wk_vector_meta_handler()
}
\arguments{
\item{x}{An object with a \code{\link[=wk_handle]{wk_handle()}} method.}
}
\value{
A data.frame with columns \code{geometry_type}, \code{has_z},
\code{has_m}, \code{srid}, and \code{size}.
}
\description{
Extract vector- and feature-level meta
}
\examples{
wk_meta(wkt("POINT (1 2)"))
wk_vector_meta(wkt("POINT (1 2)"))

}
