% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-data-frame.R, R/slice.R
\name{wk_slice.data.frame}
\alias{wk_slice.data.frame}
\alias{wk_slice}
\alias{wk_slice.default}
\title{Extract regions of handleable objects}
\usage{
\method{wk_slice}{data.frame}(handleable, from = NULL, to = NULL, ...)

wk_slice(handleable, from = NULL, to = NULL, ...)

\method{wk_slice}{default}(handleable, from = NULL, to = NULL, ...)
}
\arguments{
\item{handleable}{A geometry vector (e.g., \code{\link[=wkb]{wkb()}}, \code{\link[=wkt]{wkt()}}, \code{\link[=xy]{xy()}},
\code{\link[=rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[=wk_handle]{wk_handle()}} is defined.}

\item{from}{1-based index of the feature to start from}

\item{to}{1-based index of the feature to end at}

\item{...}{Ignored}
}
\value{
A subset of \code{handleable}
}
\description{
Extract regions of handleable objects
}
\examples{
wk_slice(xy(1:5, 1:5), from = 3, to = 5)
wk_slice(
  data.frame(let = letters[1:5], geom = xy(1:5, 1:5)),
  from = 3, to = 5
)

}
