% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{wk_void_handler}
\alias{wk_void_handler}
\alias{wk_void}
\alias{wk_debug_handler}
\alias{wk_debug}
\alias{wk_problems_handler}
\alias{wk_problems}
\alias{wkt_writer}
\alias{wkb_writer}
\alias{new_wk_handler}
\alias{is_wk_handler}
\alias{as_wk_handler}
\alias{wk_handle}
\alias{wk_handle.wk_wkb}
\alias{wk_handle.wk_wkt}
\alias{wk_handle.wk_xy}
\alias{wk_handle.wk_rct}
\title{Test handlers}
\usage{
wk_void_handler()

wk_void(x, ...)

wk_debug_handler()

wk_debug(x, ...)

wk_problems_handler()

wk_problems(x, ...)

wkt_writer(precision = 16, trim = TRUE)

wkb_writer()

new_wk_handler(handler_ptr, subclass = character())

is_wk_handler(handler)

as_wk_handler(handler, ...)

wk_handle(x, handler, ...)

\method{wk_handle}{wk_wkb}(x, handler, ...)

\method{wk_handle}{wk_wkt}(x, handler, ...)

\method{wk_handle}{wk_xy}(x, handler, ...)

\method{wk_handle}{wk_rct}(x, handler, ...)
}
\arguments{
\item{x}{A geometry vector (e.g., \code{\link[=wkb]{wkb()}} or \code{\link[=wkt]{wkt()}}).}

\item{...}{Unused}

\item{precision}{If \code{trim} is \code{TRUE}, the total number of significant digits to keep
for each result or the number of digits after the decimal place otherwise.}

\item{trim}{Use \code{FALSE} to keep trailing zeroes after the decimal place.}

\item{handler_ptr}{An external pointer to a newly created WK handler}

\item{subclass}{The handler subclass}

\item{handler}{An object created with \code{\link[=new_wk_handler]{new_wk_handler()}}.}
}
\value{
A WK handler.
}
\description{
Test handlers
}
