% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-sf.R, R/writer.R
\name{sfc_writer}
\alias{sfc_writer}
\alias{wk_writer.sfc}
\alias{wk_writer}
\alias{wk_writer.wk_wkt}
\alias{wk_writer.wk_wkb}
\alias{wk_writer.wk_xy}
\alias{wkt_writer}
\alias{wkb_writer}
\alias{xyzm_writer}
\title{Write geometry vectors}
\usage{
sfc_writer()

\method{wk_writer}{sfc}(handleable, ...)

wk_writer(handleable, ...)

\method{wk_writer}{wk_wkt}(handleable, ..., precision = 16, trim = TRUE)

\method{wk_writer}{wk_wkb}(handleable, ...)

\method{wk_writer}{wk_xy}(handleable, ...)

wkt_writer(precision = 16, trim = TRUE)

wkb_writer()

xyzm_writer()
}
\arguments{
\item{handleable}{A geometry vector (e.g., \code{\link[=wkb]{wkb()}}, \code{\link[=wkt]{wkt()}}, \code{\link[=xy]{xy()}},
\code{\link[=rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[=wk_handle]{wk_handle()}} is defined.}

\item{...}{Passed to the writer constructor.}

\item{precision}{If \code{trim} is \code{TRUE}, the total number of significant digits to keep
for each result or the number of digits after the decimal place otherwise.}

\item{trim}{Use \code{FALSE} to keep trailing zeroes after the decimal place.}
}
\value{
A \link[=wk_handle]{wk_handler}.
}
\description{
Write geometry vectors
}
