% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkb.R, R/wksxp.R
\name{wkb}
\alias{wkb}
\alias{parse_wkb}
\alias{as_wkb}
\alias{as_wkb.character}
\alias{as_wkb.wk_wkb}
\alias{as_wkb.wk_wkt}
\alias{as_wkb.wk_wksxp}
\alias{as_wkb.blob}
\alias{as_wkb.WKB}
\alias{as_wksxp.blob}
\alias{as_wksxp.WKB}
\title{Mark lists of raw vectors as well-known binary}
\usage{
wkb(x = list(), crs = wk_crs_auto())

parse_wkb(x, crs = wk_crs_auto())

as_wkb(x, ...)

\method{as_wkb}{character}(x, ..., crs = NULL)

\method{as_wkb}{wk_wkb}(x, ...)

\method{as_wkb}{wk_wkt}(x, ...)

\method{as_wkb}{wk_wksxp}(x, ...)

\method{as_wkb}{blob}(x, ..., crs = NULL)

\method{as_wkb}{WKB}(x, ..., crs = NULL)

\method{as_wksxp}{blob}(x, ..., crs = NULL)

\method{as_wksxp}{WKB}(x, ..., crs = NULL)
}
\arguments{
\item{x}{A \code{\link[=list]{list()}} of \code{\link[=raw]{raw()}} vectors or \code{NULL}.}

\item{crs}{A value to be propagated as the CRS for this vector.}

\item{...}{Unused}
}
\value{
A \code{\link[=new_wk_wkb]{new_wk_wkb()}}
}
\description{
Mark lists of raw vectors as well-known binary
}
\examples{
wkb(wkt_translate_wkb("POINT (20 10)"))

}
