% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkb.R
\name{wkb}
\alias{wkb}
\alias{parse_wkb}
\alias{wk_platform_endian}
\alias{as_wkb}
\alias{as_wkb.default}
\alias{as_wkb.character}
\alias{as_wkb.wk_wkb}
\alias{as_wkb.blob}
\alias{as_wkb.WKB}
\title{Mark lists of raw vectors as well-known binary}
\usage{
wkb(x = list(), crs = wk_crs_auto(), is_geodesic = FALSE)

parse_wkb(x, crs = wk_crs_auto())

wk_platform_endian()

as_wkb(x, ...)

\method{as_wkb}{default}(x, ...)

\method{as_wkb}{character}(x, ..., crs = NULL)

\method{as_wkb}{wk_wkb}(x, ...)

\method{as_wkb}{blob}(x, ..., crs = NULL)

\method{as_wkb}{WKB}(x, ..., crs = NULL)
}
\arguments{
\item{x}{A \code{\link[=list]{list()}} of \code{\link[=raw]{raw()}} vectors or \code{NULL}.}

\item{crs}{A value to be propagated as the CRS for this vector.}

\item{is_geodesic}{\code{TRUE} if edges must be interpolated as geodesics when
coordinates are spherical, \code{FALSE} otherwise.}

\item{...}{Unused}
}
\value{
A \code{\link[=new_wk_wkb]{new_wk_wkb()}}
}
\description{
Mark lists of raw vectors as well-known binary
}
\examples{
wkb(wkt_translate_wkb("POINT (20 10)"))

}
