% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkb.R
\name{wkb}
\alias{wkb}
\alias{as_wkb}
\alias{as_wkb.wk_wkb}
\title{Mark raw vectors as well-known binary}
\usage{
wkb(x)

as_wkb(x, ...)

\method{as_wkb}{wk_wkb}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=list]{list()}} of \code{\link[=raw]{raw()}} vectors or \code{NULL}.}

\item{...}{Unused}
}
\value{
A \code{\link[=new_wk_wkb]{new_wk_wkb()}}
}
\description{
Mark raw vectors as well-known binary
}
\examples{
wkb(wkt_translate_wkb("POINT (20 10)"))

}
