% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{wkb_coords}
\alias{wkb_coords}
\alias{wkt_coords}
\title{Extract coordinates from well-known geometries}
\usage{
wkb_coords(wkb)

wkt_coords(wkt)
}
\arguments{
\item{wkb}{A \code{list()} of \code{\link[=raw]{raw()}} vectors, such as that
returned by \code{\link[sf:st_as_binary]{sf::st_as_binary()}}.}

\item{wkt}{A character vector containing well-known text.}
}
\value{
A data.frame with columns:
\itemize{
\item \code{feature_id}: The index of the top-level feature
\item \code{nest_id}: The recursion level (if feature is a geometry collection)
\item \code{part_id}: The part index (if nested within a multi-geometry or collection)
\item \code{ring_id}: The ring index
\item \code{coordinate_id}: The coordinate index
\item \code{x}, \code{y}, \code{z}, \code{m}: Coordinaate values (both absence and \code{nan} are recorded
as \code{NA})
}
}
\description{
Extract coordinates from well-known geometries
}
\examples{
wkt_coords("LINESTRING (0 1, 19 27)")

}
