% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wksxp.R
\name{wksxp}
\alias{wksxp}
\alias{as_wksxp}
\alias{as_wksxp.default}
\alias{as_wksxp.character}
\alias{as_wksxp.wk_wksxp}
\title{Mark lists as well-known "S" expressions}
\usage{
wksxp(x = list(), crs = wk_crs_auto())

as_wksxp(x, ...)

\method{as_wksxp}{default}(x, ..., crs = NULL)

\method{as_wksxp}{character}(x, ..., crs = NULL)

\method{as_wksxp}{wk_wksxp}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=list]{list()}} features (see details)}

\item{crs}{A value to be propagated as the CRS for this vector.}

\item{...}{Unused}
}
\value{
A \code{\link[=new_wk_wksxp]{new_wk_wksxp()}}
}
\description{
Mark lists as well-known "S" expressions
}
\details{
The "wksxp" format waw an attempt to provide an R-native format
that could losslessly store WKT and WKB. This is now deprecated
and will be removed in a future version.
}
